requirements.txt
# 품질 도구
black==24.4.2
ruff==0.5.7
mypy==1.10.0
pytest==8.2.0
flake8==7.1.0
bandit==1.7.8


# FastAPI 및 관련 라이브러리
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4

# PDF 처리
PyPDF2==3.0.1
pdfminer.six==20231228
PyMuPDF==1.24.9

# OCR 처리
paddleocr==2.7.5
paddlepaddle==2.6.1

# EPUB 생성
ebooklib==0.18
lxml==5.2.2

# 이미지 처리
Pillow==10.3.0

# Supabase 연동
supabase==2.6.0
storage3==0.7.7

# 비동기 작업 큐
celery==5.3.4
redis==5.0.7
asgiref==3.7.2

# 유틸리티
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0

# 로깅
structlog==24.1.0

# 테스트
pytest-asyncio==0.23.7

# 인증/보안 (JWT, 비밀번호 해싱)
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
